requirements.txt
altgraph==0.17.2
colorama==0.4.4
future==0.18.2
pefile==2021.9.3
Pillow==8.4.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pywin32-ctypes==0.2.0
qrcode==7.3.1
